requirements.txt
aiogram==3.22.0
aiohttp==3.9.1
python-dotenv==1.0.0
psutil==5.9.8
psycopg2-binary==2.9.9 